backend/pyproject.toml
[project]
name = "research-assistant-backend"
version = "0.1.0"
description = "3D Spatial Knowledge Graph Research Assistant — FastAPI + PydanticAI Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-ai[google]>=0.1.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.5.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
